MCP Inspector 5P (2/2): extract logic panels + reduce MCPInspector to ~1200 lines
Extract three logic-bearing panels and shared types:
- chat-types.ts: ChatMessage, ExecutionRun, DisplayGroup types + groupMessages helper
- ChatPanel: chat history, ExecutionRunBlock, system prompt dialog, starter chips, input row
- ManualToolPanel: annotation badges, JsonSchemaForm, Copy JSON-RPC, Save Request dialog, ToolResult
- ServerListPanel: server cards, status/action buttons, Tools/Saved sub-tabs, tool search, saved requests list

MCPInspector.tsx reduced from ~3000 lines to ~1220 lines.
All state remains in MCPInspector; components receive via props.
